+++ modules/core/src/test/scala/org/llm4s/llmconnect/LLMConnectEnvReaderRoutingTest.scala
@@ -1,8 +1,7 @@
 package org.llm4s.llmconnect
 
-import org.llm4s.llmconnect.config.{ ContextWindowResolver, OllamaConfig, OpenAIConfig }
+import org.llm4s.llmconnect.config.{ ContextWindowResolver, OpenAIConfig }
 import org.llm4s.model.{ ModelRegistryConfig, ModelRegistryService }
-import org.llm4s.types.ProviderModelTypes.ProviderId
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 
@@ -24,19 +23,6 @@ class LLMConnectEnvReaderRoutingTest extends AnyFunSuite with Matchers {
     val res = LLMConnect.getClient(cfg)
     res match {
       case Right(client) => client.getClass.getSimpleName shouldBe "OpenRouterClient"
-      case Left(err)     => fail(s"Expected Right, got Left($err)")
-    }
-  }
-
-  test("LLMConnect.getClient returns OllamaClient for Ollama provider") {
-    val cfg = OllamaConfig.fromValues(
-      modelName = "llama3.1",
-      baseUrl = "http://localhost:11434"
-    )
-
-    val res = LLMConnect.getClient(ProviderId("ollama"), cfg)
-    res match {
-      case Right(client) => client.getClass.getSimpleName shouldBe "OllamaClient"
       case Left(err)     => fail(s"Expected Right, got Left($err)")
     }
   }